August-1
HTML (Hyper Text Markup Language) is the foundational language used to create and structure content on the web. CSS (Cascading Style Sheets) is a stylesheet language used to control the presentation of HTML elements. Bootstrap is a popular open-source front-end framework that simplifies the development of responsive and mobile-first websites.